requirements.txt
web3
azure-storage-blob
python-dotenv
